+++ src/pages/Search/components/HelpPanel/HelpPanel.module.css
@@ -10,7 +10,7 @@
 
 .selectedItemContainer {
   position: fixed;
-  right: -320px;
+  right: 0;
   top: 0;
   width: 320px;
   height: 100%;
@@ -20,12 +20,16 @@
   backdrop-filter: blur(5px);
   -webkit-backdrop-filter: blur(5px);
   opacity: 0;
+  translate3d: 100%, 0, 0;
+  transform: translate3d(100%, 0, 0);
   transition: 0.5s all cubic-bezier(0.82, 0.085, 0.395, 0.895);
 }
 
 .selectedItemContainer.showSheet {
   right: 0;
   opacity: 1;
+  translate3d: 0, 0, 0;
+  transform: translate3d(0, 0, 0);
 }
 
 .sheetTitleContainer {
